refactor(vite): move layer optimisation into separate plugin
